refactor: MainPage 첫번째 섹션 컴포넌트 분리
